+++ triggers/service/BUILD.bazel
@@ -6,6 +6,7 @@ load(
     "da_scala_binary",
     "da_scala_library",
     "da_scala_test",
+    "da_scala_test_suite",
 )
 load("@build_environment//:configuration.bzl", "sdk_version")
 load("@os_info//:os_info.bzl", "is_windows")
@@ -82,8 +83,8 @@ da_scala_binary(
     ],
 )
 
-da_scala_test(
-    name = "tests",
+da_scala_library(
+    name = "trigger-service-tests",
     srcs = glob(["src/test/scala/com/digitalasset/daml/lf/engine/trigger/*.scala"]),
     data = [
         ":test-model.dar",
@@ -104,7 +105,6 @@ da_scala_test(
         "//language-support/scala/bindings-akka",
         "//ledger-api/rs-grpc-bridge",
         "//ledger-api/testing-utils",
-        "//ledger-service/cli-opts",
         "//ledger-service/jwt",
         "//ledger/caching",
         "//ledger/ledger-api-auth",
@@ -115,7 +115,6 @@ da_scala_test(
         "//ledger/participant-state",
         "//ledger/sandbox-classic",
         "//ledger/sandbox-common",
-        "//libs-scala/flyway-testing",
         "//libs-scala/ports",
         "//libs-scala/postgresql-testing",
         "//libs-scala/resources",
@@ -129,7 +128,7 @@ da_scala_test(
         "@maven//:com_typesafe_akka_akka_parsing_2_12",
         "@maven//:eu_rekawek_toxiproxy_toxiproxy_java_2_1_3",
         "@maven//:io_spray_spray_json_2_12",
-        "@maven//:org_flywaydb_flyway_core",
+        "@maven//:org_scalactic_scalactic_2_12",
         "@maven//:org_scalatest_scalatest_2_12",
         "@maven//:org_scalaz_scalaz_core_2_12",
     ],
@@ -156,6 +155,39 @@ da_scala_test(
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_typesafe_akka_akka_http_core_2_12",
         "@maven//:io_spray_spray_json_2_12",
+        "@maven//:org_scalatest_scalatest_2_12",
+        "@maven//:org_scalaz_scalaz_core_2_12",
+    ],
+)
+
+da_scala_test_suite(
+    name = "test",
+    srcs = glob(["src/test-suite/scala/**/*.scala"]),
+    deps = [
+        ":trigger-service",
+        ":trigger-service-tests",
+        "//daml-lf/archive:daml_lf_archive_reader",
+        "//daml-lf/archive:daml_lf_dev_archive_java_proto",
+        "//daml-lf/data",
+        "//language-support/scala/bindings-akka",
+        "//ledger-api/rs-grpc-bridge",
+        "//ledger-api/testing-utils",
+        "//ledger-service/cli-opts",
+        "//ledger-service/jwt",
+        "//ledger/ledger-api-auth",
+        "//ledger/ledger-api-common",
+        "//ledger/ledger-resources",
+        "//ledger/sandbox-classic",
+        "//ledger/sandbox-common",
+        "//libs-scala/flyway-testing",
+        "//libs-scala/ports",
+        "//libs-scala/postgresql-testing",
+        "//libs-scala/resources",
+        "//triggers/service/auth:oauth-test-server",
+        "@maven//:com_typesafe_akka_akka_http_core_2_12",
+        "@maven//:eu_rekawek_toxiproxy_toxiproxy_java_2_1_3",
+        "@maven//:io_spray_spray_json_2_12",
+        "@maven//:org_flywaydb_flyway_core",
         "@maven//:org_scalatest_scalatest_2_12",
         "@maven//:org_scalaz_scalaz_core_2_12",
     ],
