+++ triggers/service/src/test-suite/scala/com/daml/lf/engine/trigger/ServiceConfigTest.scala
