+++ triggers/service/src/test-suite/scala/com/daml/lf/engine/trigger/TriggerServiceTestWithDb.scala
@@ -1,0 +1,9 @@
+// Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.lf.engine.trigger
+
+class TriggerServiceTestWithDb
+    extends AbstractTriggerServiceTest
+    with AbstractTriggerServiceTestWithDb
+    with AbstractTriggerServiceTestNoAuth {}
